% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{next_x}
\alias{next_x}
\title{grab the next x value}
\usage{
next_x(x1, numPoints, noise, y1, maxLength, curliness)
}
\arguments{
\item{x1}{the starting x value}

\item{noise}{the perlin noise matrix generated by the create() function}

\item{y1}{the y value associated with the starting x value}

\item{maxLength}{the largest possible length of a line segment}

\item{curliness}{the multiplier to determine the angle to direct the line segment}
}
\description{
generates the next points to be used to finish each line segment in the image
}
\details{
uses perlin noise and a multiplier to generate a segment's endpoint
}
