% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{next_y}
\alias{next_y}
\title{grab the next y value}
\usage{
next_y(y1, numPoints, noise, x1, maxLength, curliness)
}
\arguments{
\item{y1}{the starting y value}

\item{noise}{the perlin noise matrix generated by the create() function}

\item{x1}{the x value associated with the starting y value}

\item{maxLength}{the largest possible length of a line segment}

\item{curliness}{the multiplier to determine the angle to direct the line segment}
}
\description{
generates the next points to be used to finish each line segment in the image
}
\details{
uses perlin noise and a multiplier to generate a segment's endpoint
}
